+++ deps.ts
@@ -2,5 +2,7 @@ export {
   css,
   ResinCssEmitter,
   ResinCssGlobalStyle,
-} from 'https://deno.land/x/resin@v0.1.6/mod.ts';
+} from 'https://deno.land/x/resin@v0.1.7/mod.ts';
 export * as classNames from 'https://deno.land/x/classnames@0.1.1/index.ts';
+export type { Keystroke } from 'jsr:@carcajada/teclas@^1.0.8';
+export { handleKeyboard } from 'jsr:@carcajada/teclas@^1.0.8';